requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
sqlalchemy
authx
psycopg2
passlib
pydantic[email]